✨ feat: implement API-first architecture for burncloud-cli
- Add reqwest, serde, and dirs dependencies to crates/cli.
- Implement ApiClient in src/client.rs for token management and remote API interaction.
- Transition CLI from local core management to a headless API client suitable for automation.

+++ crates/service/crates/service-monitor/src/collectors/disk.rs
@@ -129,9 +129,9 @@ impl DiskCollector {
                 )));
             }
 
-            let block_size = statvfs.f_frsize as u64;
-            let total_blocks = statvfs.f_blocks as u64;
-            let free_blocks = statvfs.f_bavail as u64;
+            let block_size = statvfs.f_frsize;
+            let total_blocks = statvfs.f_blocks;
+            let free_blocks = statvfs.f_bavail;
 
             let total = total_blocks * block_size;
             let available = free_blocks * block_size;
